backend/requirements.txt
# RENDER-COMPATIBLE REQUIREMENTS (Python 3.11)
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Essential dependencies with compatible versions
numpy==1.25.2
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
aiofiles==23.2.1

# Utilities
jinja2==3.1.2

# ML dependencies (if needed - add back gradually)
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4
# pandas==2.1.4
